Added a custom update dialog + bug fixes.
- Added a window that asks the user and asks if they want to the see the
online patch notes.
- Made the GUI and command line version have different configs and they
are both stored in the data folder.
- Added the method of creating the log folder in the command line
version, so the program would not become unstable.
- Added the licenses for all the open source libraries this program
uses.
- Fixed linux shebang line to the correct version of python. (In the
command line version.)
The bug fixes include:
- When a messagebox appeared, a blank tk window would show up. (Fixed by
creating a blank tk window when the program starts, but it is
withdrawn.)
- The license window would not fit all of the text on the screen. (Made
it a little bigger.)
-  The license window text box would cut off at a certain point when the
window was extended. (Made the textbox the window size.)