Updated dead links, error handling, slightly more user friendly
made Converter slightly less annoying by it not firing a ton of error message boxes on first startup.
Also when saving as a text file, the program appends .txt to the file name which you can change in the settings wHEN I REwRITE the UI.

updated dead update urls to new urls.

changed the way program handles errors during conversion.
e.g when the program incounters a problem, instead of running in the background forever, it displays the main conversion window again.

0.2a released latest by end of next week? maybe? hopefully.